Rebuild cache reliably when upgrading transitive dependencies
Fix https://github.com/elm/compiler/issues/1710

Before I decided if the summary was valid based on a solution.dat file.
It contained a (Map Pkg.Name Pkg.Version) of the exact packages
represented in the summary.

That was not enough information though. The ExposedModules in the
Summary need to change if something moves from
“transitive-dependencies” into “dependencies”.

So the new version stores the full description of the dependencies
right in the summary.dat file. The idea is that:

- If it is valid, you need to read the whole file anyway.
- If it is invalid, you have to read a bunch of files in packages
anyway.

So the common path (if it is valid) should be one file read faster, and
the uncommon path (if it is invalid) should be one large file read
slower.