Remove Elm.Name and switch elm.cabal file to new name format
Compared to the 0.19 release when compiling elm-spa example from
scratch on my computer, this does 17% less allocation and total time
goes from 0.45 to 0.38 (~15%)

Being able to customize the binary serialization format also means that
elmi, elmo, and dat files are a decent bit smaller. No more using
Word64 for the length of variables like “a” and “test”

Need to do some testing with larger codebases to see how this extends
though.