Get rid of Generate.Artifacts
In the old implementation, all artifacts were gathered first. From
there, they would be written or ignored depending on whether this was a
project that needs them or a dependency that does not.

This has a very tricky downside. Haskell is lazy, so in the old
implementation, waiting to write the artifacts also delayed the
creation of these artifacts. That meant that computations that SHOULD
be happening concurrently on a bunch of light-weight threads were
actually escaping and running sequentially at a later time.

(The amount of time lost on performance problems due to laziness is
immense both for developers and users, but hey, on the other side of
the scale, you can write [0..] instead of [0 .. length xs - 1] when you
need an equivalent of List.indexedMap! ¯\_(ツ)_/¯)

Anyway, the new implementation will write the artifacts on each
light-weight thread, forcing evaluation such that the computations will
actually run concurrently on different processors.