Move Deps.Cache to cleaner Deps.Registry module
The new module provides more focused functions, allowing other parts of
the compiler to decide if they want to use an Http.Manager or not. If
they decide to use it, they can then thread it around themselves. This
moves that decision to specific endpoints like `elm make` and `elm
repl` allowing them to speed things up on a case-by-case basis.