Collapse Deps.Explorer into Deps.Solver with new CPS implementation
Get rid of dependency on LogicT by creating my own version. Doing it
this way lets me track "deadends" in a way that I hadn't realized was
possible before. This may make it possible to give nicer hints when
something is not solving, but ultimately, I want to get folks using a
UI to mess with their dependencies.