Update interfaces to be more precise in new compilation scheme
There is now a distinction between public and private interfaces. This
will serve as an explicit reminder that it is necessary to keep the
type information around for the debugger.

So these changes are also the foundation of the fix for the Map.!
errors caused when the debugger is trying to gather all the
transitively used types.

Also adds Eq to Interface which will allow the comparison against
cached interfaces. If things are the same, subsequent compiles can be
skipped.