Revamp compilation process
The old code went directly from a ByteString to the compile artifacts.
This had the downsides of:

1. Parsing had to happen along with the rest of the process. So we had
to parse the header to crawl, and then parse the header again when
compiling.
2. It always constructed the interface for the file. This is necessary
for any `elm make` call, but it is not needed for `elm repl`. In the
REPL, it actually makes things a bit trickier since all you really want
is the objects and the type of the most recently evaluated expression.

So the new approach has you give a Src.Module and gives a less
processed set of artifacts. The code has also been cleaned up a bit.