Handle imports more nicely during canonicalization
Changes include:

1. The old approach had a mapping from Raw to Canonical names, so you
could then look up the canonical names in the dictionary of interfaces.
The new approach just has a mapping from Raw names to interfaces which
contain the relevant package information.

2. There used to be an intermediate Import type. There was code for
checking if all imports existed. This check happens in the crawl phase
though, so you could never get the error to manifest in this spot. The
new approach just gets the necessary information without an
intermediate allocation. (The allocation is so small it shouldn't
matter though. More about it being checked elsewhere.)