Redo Deps.Verify code in Elm.Details
The new code is similar, but the process of downloading and verifying
that all the packages build together is changed.

The new code runs all the package verification in threads. Instead of
being a multiphase process that crawls and then builds, the code here
does it all in one pass. This relies on modules being checked for
cycles on publication.

The current design is that there are three cached files that are useful
at different times:

1. d.dat stores the details. So it is loaded on any `elm make`
invocation. It stores timestamps and metadata for elm.json and Elm
files, making it possible to skip work and file system calls in more
cases. It also indicates if the other files are fresh. This file will
be read and written on each compile, so I have tried to keep it pretty
small.
2. i.dat stores all of the foreign interfaces that you might need in
the canonicalization phase.
3. o.dat stores the object graph for all transitive dependencies. You
need this if your program gets to the generation phase. Someday, maybe
it can be topologically sorted so that you only read into memory the
parts you actually need.

Might be worthwhile to combine i.dat and o.dat into one file to reduce
number of system calls, but then folks with packages or with
--output=/dev/null pay the price. But as is, it should make compilation
a bit faster if you add the --output=/dev/null flag. Maybe that is more
valuable.

NOTE: it may be necessary to limit the number of concurrent downloads!