Update kernel parsing to work properly
Needed to make a couple changes:

- Make sure all strings are copied out of the file strictly. Otherwise
you get segmentation faults when trying to read the memory later.
- Elm fields are counted later when added to the global object graph
- When an import is not found, assume it is from the local package. If
the import really does not exist, that will be found by whatever is
running the kernel parser. That is why we give out the imports directly
now. (This should also protect against cases where a dependency that
goes through a kernel file is not captured by other imports.)