Add new build manager that should allow more concurrency
In the old system, it was possible that computations would escape from
threads because the values were evaluated lazily. So instead of doing
the work on all the threads, the work was put in a thunk and passed
along.

In the new setup, I was careful to make sure that the interfaces and
object files are forced on the threads directly. The objects are
written immediately, forcing that part. With the interfaces, I read the
old .elmi file and do a comparison. If they are fully equal, I know
that the computation has been fully forced, and that I do not need to
recompile files that depend on this since the interface has not
changed. This should be helpful in reducing incremental compile times,
where many edits are changing little details rather than changing the
API.

There are still many aspects that are incomplete in this
implementation, so more work is needed to make sure the concurrency
gives worthwhile performance improvements now that laziness is PROBABLY
not messing it up.