Generate.JavaScript.generate just gives a Builder now
The decision whether to generate JS used to be spread across a couple
different modules, some information here, calling some function there,
putting all back together.

The new setup means that decision is made in Make.hs, keeping track of
how many mains there are and the implications of that. It calls this
function only when all that is good and it definitely wants a Builder.