Have a specific abstraction for tracking types in --debug
There is a bug in 0.19.0 in which --debug can fail if certain types are
used in messages, which need to be encoded into the signature of the
resulting debugger.

By making a specific data structure, hopefully it'll be more obvious
that private type information cannot be thrown away, even when it is in
dependencies of dependencies!