Revamp solver
Add ability to backtrack and to error. This means errors never get
skipped over.

Also add initEnv that sets up the Http.Manager and figures out if it
can connect to package.elm-lang.org. This way different entrypoints can
all use the same logic to know if they are trying to work with packages
in online/offline modes.