Start to track context in errors only
This has two benefits:

- Before there were two pieces of information (Context and Expectation)
and it made it difficult to make nice error messages. Almost all
expectations are contextual, but there was no connection between these
two values at the type level. So in the new version, the error types
will be nested, thereby tracking the full context in the error itself.

- It means the parse state does not need to allocate a pointer to the
context anymore. Perhaps this improvement will offset any costs that
are incurred in the new set up.