No longer need an Elm.Compiler module as interface into other code
This was more of a passthrough to the real code. It made it easier to
have a "public interface" between the compiler part of the code and the
various binaries that used it. Even before the code is moved to a
monolith, this mainly added a layer of indirection that was confusing.