Introduce OneOrMore.getFirstTwo for one specific scenario
There is one line of code in Canonicalize.Environment.Dups where you
want to see only one version, and otherwise, you have a guarantee that
you have two or more values.

This function is only used in the case of an error, thereby avoiding
using OneOrMore.destruct which would force some allocations in the
normal case.

And it is one fewer "incomplete pattern" which is nice.