Make it clearer that the suggestions are not all possible modules
It is only able to suggest modules that it has seen before, or modules
it happened to load during the current build. If it is a typo, there
may actually be a file there with the right name, but it wouldn't get
suggested if no one had ever referred to it by that name.