Introduce concept of parse snippets
This is necessary to delay the parsing of documentation comments. It is
very rare that you want to turn comments into JSON, so rather than
doing the allocations on every run, the allocations are delayed until
it's clear that they are needed.

This requires holding onto the (ForeignPtr Word8) of the initial
ByteString so that the memory region stays alive that far into
compilation. As a result, pretty much all the parse modules will need
slight alterations to add the extra variable again.

This change also gives a hand-written definition of (<*>) that
hopefully is a bit better than one derived from a Monad instance or
whatever.