Introduce many distinct "string" types to help ensure appropriate escaping on code gen
I initially made the new (Utf8 size tipe) type with the idea that the
crucial distinction was how the Binary code is generated (with a Word8
length or a Word64 length) and that there was some value in
distinguishing types of strings a bit more than that.

The new (Utf8 tipe) gets rid of the concept of size. That module now
serves as a collection of helpers that are exposed through modules like
Elm.String and Json.String.

In this world, each particular (Utf8 ELM_STRING) or (Utf8 JSON_STRING)
decides individually if it should have a binary instance at all. In the
case of JSON_STRING, there shouldn't be one. If it is getting
serialized, it is probably a bug or a mistake!

Similarly, I can expose a really minimal API for each individual type
of String. In cases like (Utf8 ELM_NAME) you know it's safe to generate
them directly based on parser invariants. No need to escape anything.
Whereas with turning a documentation comment into a (Utf8 JSON_STRING)
you actually need to escape newlines, so allocation needs some extra
checks and logic.

Point is, the new setup should make it harder to make mistakes! (It
actually caught one in Elm.Package by distinguishing AUTHOR and
PROJECT. That'll be checked in later though!)