update Json.Decode to incorporate parse snippets and Json.String
One of the crucial changes is the introduction of customString which
lets you define a decoder based on a parser! This means JSON strings
containing versions, package names, module names, types, etc. can
directly use the existing parsing infrastructure for validation with
minimal allocation.

This technique is also used in the new KeyDecoder idea. There are a
couple dictionaries in elm.json and various JSON data where the keys
are package names or something with distinct structure. This makes it
possible to use parsers for that directly, cleaning up some code and
reducing allocation a bit.

JSON decoding is not a big cost though. The main point is just having a
coherent system for handling Json.String escaping reliably.