use smaller Elm.License representation / remove binary instance
This information is not needed at runtime, and there is no reason to
serialize it to disk. It only needs to be verified in elm.json and kept
around long enough to generate a new elm.json if necessary.

The new API reduces the surface area to try to keep it to that.