start using D.customString in Elm.Package
This also introduces Pkg.keyDecoder for decoding constraints and
versions in elm.json

I also moved some code for finding nearby names into this module since
the AUTHOR and PROJECT types are internal to the module. This actually
revealed a bug in which the check for elm and elm_explorations was
actually happening in the projectDistance function, which should never
happen!