Just use ByteStrings in Elm.Kernel instead of some Utf8 type
This means the o.dat files will stick around in pinned memory longer.

It is only needed for code gen, so we can avoid loading it for files
without main or if someone is using --output=/dev/null

But in cases where it's needed, this means that the o.dat memory region
will live for the duration of code gen. It seems okay though. In an
ideal world, we could go directly from pinned memory to the generated
file without any intermediate allocation, so it feels like it's in an
okay direction.