Add in more extensive escaping for GLSL shaders
Fixes https://github.com/elm/compiler/issues/1923

This should allocate a bit less during parsing by skipping a (Utf8
SHADER_STRING) and going directly to [Char] which is needed for
Language.GLSL.Parser.

It should then be stored as a ByteString which will keep *.elmo files
alive longer during code gen. It means there is less memory traffic
that relates to GC, but at the cost of using pinned memory for longer.
Again, I think this is okay / in a desirable direction.