introduce DocsGoal GADT to help generate docs in package projects
This relies on a function called finalizeDocs that has the coolest type
I have seen in a while. Did not realize GADTs could do this, where each
branch actually has a "different" return type, but it is still
guaranteed to be used in a reasonable way because of the type
variables. Really neat! So you can accomplish the same thing as giving
a callback, but without giving a callback!