remove waiting gif in reactor
I think the new style gives HTML with JS inline. That means the compile
will be done by the time the browser gets the HTML, so the waiting
animation wouldn't show.

Would need to structure it differently to show a waiting animation, but
hopefully with the incremental compiles, this wouldn't really make
sense to have.

Maybe it makes sense to log package downloads on the terminal though.
That seems like something that'd make a slower build.