fix bugs in fromTypeVariable and fromTypeVariableScheme
# PROBLEM

The problem started with the fact that getIndexSize was initially
written to already include the +2 needed by fromVarIndex. I guess I
hand optimized that, but did not reflect it in the name clearly.

I guess after that, I used the getIndexSize function in
fromTypeVariable and fromTypeVariableScheme without accounting for that
padded +2. That meant they both were producing strings with 2 spots of
whatever from the memory that happens to be around when the MBA is
allocated.

# STORY

We were compiling a project with 700+ modules. On the initial build,
all the files were compiled. With the new compiler, we would expect
touching one file to result in only one file being compiled, no matter
how many dependencies it has. (It compares against the public interface
from last time, and if it's the same, we know everything will keep
working in all the dependencies!)

So on the incremental build, we touch a file that is imported by many
other files, but instead of 1 file being compiled, we see 108. Weird!
Trying this again on a second incremental build, we saw just 1 file
being compiled. Very weird!

It turned out that if you exposed values without type annotations, it
was possible that the compiler would need to generate some type
variable names for you. Those would then be written into the interface
files, and subsequently compared to other interface files to see if it
is exactly the same. So with this bug unresolved, it was possible for
the messed up names to clash in this context!