use alternate strategy for creating fake names
Used for top-level cyclic data and for destructuring patterns. In both
cases, the restriction on name shadowing should be enough insurance to
make sure this strategy of creating fake names works alright.

The important thing is that this does not produce names that are likely
to be 256 chars or longer.