produce a value for (fromManyNames [])
It seems that the compiler has been handling this case for a while. I
believe the code would produce the name "" on _ patterns in the past,
and that name was thrown out at a later phase in compilation somehow.

I suspect the function should be (fromManyNames : N -> [N] -> N) to
guarantee that at least one name is given, but I do not want to get
into the implications of that change at this point in time.