make REPL behavior more predictable for multi-line entries
Introduces the Lines type to keep track of all the lines that have been
input so far in an explicit way. (Before, newlines were getting added
kind of haphazardly.)

From there, introduce ifDone and ifFail to enforce the following rules:

- If a single-line entry parses, evaluate it.
- If a single-line entry does not parse, let people give a second line.
- If a multi-line entry ends with a blank line, evaluate it.
- If a multi-line entry does not end with a blank line, give another
line.

This should help with some of the cases that @rlefevre shared:

> case True of
|   True -> "hi"
XXX

Now it will be:

> case True of
|    True -> "hi"
|    False -> "bye"
|
"hi" : String