start with def name when the first REPL lines are a type annotation
Say you type:

> incr : Int -> Int

Now when you press enter, it is going to prefill the next line like
this:

> incr : Int -> Int
| incr

This should be nice in most cases.

If someone wants a multi-line type annotation, they can either (1) end
it with a -> to make sure the type annotation appears incomplete or (2)
just press backspace and keep going with their type.