thread BuildID logic through Build.hs
As explained in the NOTE on Details.Local, it is necessary to track the
BuildID of the last time the interface changed and the last time the
file was compiled. This makes it possible to verify that all files have
been compiled AFTER the most recent interface change of their imports
in a project with multiple entrypoints such that some files are
compiled inconsistently.

Fixes https://github.com/elm/compiler/issues/1956