always show three lines of JSON in syntax error messages
Some specific errors benefit from having the line above, like in
https://github.com/elm/error-message-catalog/issues/264

Since the JSON syntax errors are always for a single position (no
context stack!) I figured I did a fix that just includes two extra
lines for all JSON syntax errors. This should not be too disruptive and
should give more of a feeling of place within the JSON file, especially
when it's a single line with something weird on it!