provide draft of worker.elm-lang.org/repl endpoint
The goal is to make it possible to simulate `elm repl` from within
guide.elm-lang.org such that people can get quite a ways into the guide
before installing anything.

The hope is that by the time anyone is running into install issues with
PATH or editors, they are already having a nice time with Elm and
interested in investing more time in it.