switch to smaller/safer representation for canonical env
Gets rid of the Data.Map.Internal usage that was used to allocate less.
The replacement is an actual model of the data that we are working
with, capturing the fact that "empty dictionary" is not a possibility.
It also throws away some data when things are ambiguous, and it should
allocate less overall.