remove draft of Endpoint.Donate for now
The /donate endpoint never went live, but it may return at a later
time. This is mostly about removing unfinished code so that it does not
go live by accident.

The intent of such an endpoint was so that Elm Software Foundation
could begin to take donations to support in-person events. Maybe that
would be conference grants, events like Elm Bridge, etc. The work
happened around Jan/Feb 2020 and was paused when the COVID 19 outbreak
began. With lockdowns, travel restrictions, and event cancelations, it
did not seem to be a good time to pursue funds for in-person community
events.

Looking forward to a time when such events are more straight-forward to
plan and attend.